requirements.txt
fastapi
uvicorn
line-bot-sdk>=3.0.0
yfinance
google-generativeai
python-dotenv
aiosqlite
pydantic
